Move diacritic Backspaces into function.
When adding diacritics for omicron and epsilon, there will need to be the option to do nothing (e.g., these vowels don't conventionally take a circumflex, so if the user tries to add one, we don't want anything to happen: previous behavior would have sent a Backspace regardless). Rather than dealing with complicated previous-state logic and reverting the status variables to prior values, it is much simpler just to just move Backspaces into the function and not send them if we don't want to do anything.